small type mismacth error fix

+++ UrGuide.Model/Posts/PostModel.cs
@@ -32,7 +32,7 @@ namespace UrGuide.Model.Posts
         public string StartTime { get; set; }
         public bool HasReserved { get; set; }
         public bool HasReacted { get; set; }
-        public short ReactionType { get; set; }
+        public int ReactionType { get; set; }
         public int BidCount { get; set; }
         public int ItineraryCount { get; set; }
         public ICollection<string> Categories { get; protected set; }
